feat(week-view): drop Manage-speakers link, add per-row chat launcher
The week view's Speakers section had a top-level "Manage speakers"
link that opened the AssignDialog — duplicating an entry point
that already exists on the schedule view's Sunday card. Removed
both the link and the dialog wiring; speakers are still added /
edited from the schedule.

Each speaker row now hosts its own chat icon (the same one used on
the schedule), so a bishop can jump straight into the conversation
for a single speaker. The chat launcher (icon + invitation
subscription + dialog + URL deep-link auto-open) is extracted to
SpeakerChatLauncher so both the schedule's SpeakerRow and the
week view's SpeakerListRow share one implementation.
