+++ src/features/meetings/sections/SpeakersSection.tsx
@@ -1,10 +1,7 @@
-import { useMemo, useRef, useState } from "react";
+import { useMemo, useState } from "react";
 import type { WithId } from "@/hooks/_sub";
 import type { MidItem as MidItemType, NonMeetingSunday, Speaker } from "@/lib/types";
 import { reorderSpeakers } from "@/features/speakers/speakerActions";
-import { AssignDialog } from "@/features/schedule/AssignDialog";
-import { SpeakerEditList, type SpeakerEditListHandle } from "@/features/schedule/SpeakerEditList";
-import { formatShortDate } from "@/features/schedule/dateFormat";
 import { ProgramSection } from "../program/ProgramSection";
 import { updateMeetingField } from "../updateMeeting";
 import { MidPlaceholderRow, SpeakerListRow } from "./SpeakerListRow";
@@ -24,9 +21,6 @@ export function SpeakersSection({ wardId, date, speakers, mid, nonMeetingSundays
   const items = useMemo(() => buildItems(ordered, mid, midLabel), [ordered, mid, midLabel]);
   const [dragIdx, setDragIdx] = useState<number | null>(null);
   const [overIdx, setOverIdx] = useState<number | null>(null);
-  const [manageOpen, setManageOpen] = useState(false);
-  const [saving, setSaving] = useState(false);
-  const editListRef = useRef<SpeakerEditListHandle>(null);
 
   function persist(nextItems: Item[]) {
     const speakerIds = nextItems.filter((x) => x.kind === "speaker").map((x) => x.id);
@@ -53,51 +47,12 @@ export function SpeakersSection({ wardId, date, speakers, mid, nonMeetingSundays
     setOverIdx(null);
   }
 
-  async function handleManageSave() {
-    if (!editListRef.current) return;
-    setSaving(true);
-    try {
-      await editListRef.current.save();
-      setManageOpen(false);
-    } finally {
-      setSaving(false);
-    }
-  }
-
-  const manageButton = (
-    <button
-      type="button"
-      onClick={() => setManageOpen(true)}
-      className="ml-auto font-sans text-[12.5px] font-semibold text-bordeaux hover:text-bordeaux-deep hover:underline hover:underline-offset-2 transition-colors"
-    >
-      Manage speakers →
-    </button>
-  );
-
-  const dialog = (
-    <AssignDialog
-      open={manageOpen}
-      title={formatShortDate(date)}
-      saving={saving}
-      onClose={() => setManageOpen(false)}
-      onSave={() => void handleManageSave()}
-    >
-      <SpeakerEditList
-        ref={editListRef}
-        date={date}
-        wardId={wardId}
-        nonMeetingSundays={nonMeetingSundays}
-      />
-    </AssignDialog>
-  );
-
   if (items.length === 0) {
     return (
-      <ProgramSection id="sec-speakers" label="Speakers" rightSlot={manageButton}>
+      <ProgramSection id="sec-speakers" label="Speakers">
         <p className="font-serif italic text-[13.5px] text-walnut-3 py-1">
-          No speakers yet. Use <em>Manage speakers</em> to add them.
+          No speakers yet. Add them from the schedule view.
         </p>
-        {dialog}
       </ProgramSection>
     );
   }
@@ -109,9 +64,8 @@ export function SpeakersSection({ wardId, date, speakers, mid, nonMeetingSundays
       label="Speakers"
       count={ordered.length}
       rightSlot={
-        <span className="ml-auto inline-flex gap-3 items-center">
-          <span className="font-serif italic text-[12.5px] text-walnut-3">Drag to reorder</span>
-          {manageButton}
+        <span className="ml-auto font-serif italic text-[12.5px] text-walnut-3">
+          Drag to reorder
         </span>
       }
     >
@@ -138,12 +92,13 @@ export function SpeakersSection({ wardId, date, speakers, mid, nonMeetingSundays
               speaker={item.speaker}
               index={speakerCounter++}
               isLast={isLast}
+              wardId={wardId}
+              date={date}
               {...drag}
             />
           );
         })}
       </ul>
-      {dialog}
     </ProgramSection>
   );
 }
